add : 로그의 회원가입 기능구현 api 통신완료
